My About me Page using HTML and CSS
This is my About Me page using Html and CSS to style my website. Throughout the code I used multiple <div class=""> and <section class=""> to be able to format and style the website the way i wanted it to in styles.css. I used a main container to select all the selection and div tags and gave it simple styling to simplify the websites styling. I used flexboxes to format the websites layout the way I wanted to, using parent and child elements (container, selectors, and divs). The website includes my background, interests, and goals I have for the future. The website also includes a youtube video by using the iframe tag. I was not able to center the video in styles.css, therefore I simply used the <center> tag to get the result I was looking for.